refactor(edu): rename package
